+++ test/libsolutil/IpfsHash.cpp
@@ -29,7 +29,7 @@ using namespace std;
 namespace solidity::util::test
 {
 
-BOOST_AUTO_TEST_SUITE(IpfsHash)
+BOOST_AUTO_TEST_SUITE(IpfsHash, *boost::unit_test::label("nooptions"))
 
 BOOST_AUTO_TEST_CASE(test_small)
 {