+++ test/yulPhaser/Phaser.cpp
@@ -103,7 +103,7 @@ protected:
 	};
 };
 
-BOOST_AUTO_TEST_SUITE(Phaser)
+BOOST_AUTO_TEST_SUITE(Phaser, *boost::unit_test::label("nooptions"))
 BOOST_AUTO_TEST_SUITE(PhaserTest)
 BOOST_AUTO_TEST_SUITE(GeneticAlgorithmFactoryTest)