+++ test/yulPhaser/Selections.cpp
@@ -34,7 +34,7 @@ using namespace solidity::util;
 namespace solidity::phaser::test
 {
 
-BOOST_AUTO_TEST_SUITE(Phaser)
+BOOST_AUTO_TEST_SUITE(Phaser, *boost::unit_test::label("nooptions"))
 BOOST_AUTO_TEST_SUITE(SelectionsTest)
 BOOST_AUTO_TEST_SUITE(RangeSelectionTest)